[Modify] 로그아웃 공통 부분 모달로 통합
